Add bash_profile, better sourcing and clear alias
Added the bash_profile file, which does not thing except sources the
bashrc file. This is set up like this because depending on if you use
linux vs mac one requires bash_profile and the other requires bashrc so
having both makes it easy to work with.

Added a source call for the zshrc to source the bashrc on load which in
turn sources the bash_profile

Added an alias to clear the screen and print my "WELCOME MR. MROCZKA"
message